requirements.txt
pypdf
langchain_community
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
huggingface_hub
fastapi
python-multipart
uvicorn
openai
tiktoken
unstructured
python-dotenv
langchain-huggingface
langchain-openai
google-cloud-storage
cloud-sql-python-connector==1.9.2
fastapi==0.111.0
google-auth==2.29.0
pg8000==1.31.2
SQLAlchemy==2.0.30
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.3.0